requirements.txt
tabulate[widechars]~=0.8.7
keyring~=22.0.1
requests~=2.23.0
rsa~=4.6
beautifulsoup4~=4.9.3
python-dateutil~=2.8.1